// src/modules/ecommerce/edit-product/widgets/basic-information/types.ts
export type BasicInformationData = {
  title: string;
  handle: string;
  shortDescription: string;
  description: string;
};

export type BasicInformationDictionary = {
  title: string;
  productTitle: string;
  productTitleRequired: string;
  urlHandle: string;
  urlPrefix: string;
  shortDescription: string;
  characters: string;
  description: string;
  loading: string;
  error: string;
};